Add CreditMemo support to QBXML
I wrote these changes originally in 2010 and shared with Keith then in 2012 to
to merge in. AFAIK, he has done that - a quick check looked like credit memo
support is present, but I'm not in a position to easily and quickly test it, as
the library's APIs have changed significantly.

Looks like my changes were mostly as follows, all around Web Connector / QBXML
1. Add support for Credit Memos (Add/Query operations)
2. Add support for CheckMod
3. Add support for ReceivePayment Mod/Query
4. Few tweaks to some of the QBXML objects where I noticed issues when I was
   using them. Unfortunately, didn't provide a lot of clear notes to myself
   why/what problem I saw that prompted me to make the change.